Introduce generic actions cell editor
