+++ renderer/ClienteActionsCellEditor.java
@@ -1,22 +1,14 @@
 package com.pinguela.rentexpres.desktop.renderer;
 
-import static com.pinguela.rentexpres.desktop.util.AppIcons.DELETE;
-import static com.pinguela.rentexpres.desktop.util.AppIcons.EDIT;
-import static com.pinguela.rentexpres.desktop.util.AppIcons.VIEW;
-
 import java.awt.Component;
-import java.awt.FlowLayout;
 import java.awt.Frame;
 import java.util.EventObject;
 import java.util.function.Supplier;
 
-import javax.swing.AbstractCellEditor;
-import javax.swing.ImageIcon;
-import javax.swing.JButton;
 import javax.swing.JOptionPane;
-import javax.swing.JPanel;
 import javax.swing.JTable;
-import javax.swing.table.TableCellEditor;
+
+import com.pinguela.rentexpres.desktop.renderer.AbstractActionsCellEditor;
 
 import com.pinguela.rentexpres.desktop.dialog.ClienteDetailDialog;
 import com.pinguela.rentexpres.desktop.dialog.ClienteEditDialog;
@@ -27,13 +19,8 @@ import com.pinguela.rentexpres.exception.RentexpresException;
 import com.pinguela.rentexpres.model.ClienteDTO;
 import com.pinguela.rentexpres.service.ClienteService;
 
-public class ClienteActionsCellEditor extends AbstractCellEditor implements TableCellEditor {
+public class ClienteActionsCellEditor extends AbstractActionsCellEditor {
 	private static final long serialVersionUID = 1L;
-
-	private final JPanel panel = new JPanel(new FlowLayout(FlowLayout.CENTER, 6, 0));
-	private final JButton btnView = iconButton(VIEW, "Ver");
-	private final JButton btnEdit = iconButton(EDIT, "Editar");
-	private final JButton btnDel = iconButton(DELETE, "Borrar");
 
         private final Frame frame;
         private final ClienteService service;
@@ -43,13 +30,11 @@ public class ClienteActionsCellEditor extends AbstractCellEditor implements Tabl
 
         public ClienteActionsCellEditor(Frame owner, ClienteService service, ActionCallback reload,
                         Supplier<ClienteDTO> rowSupplier) {
+                super();
                 this.frame = owner;
                 this.service = service;
                 this.reload = reload;
                 this.rowSupplier = rowSupplier;
-                panel.add(btnView);
-                panel.add(btnEdit);
-                panel.add(btnDel);
 
 		btnView.addActionListener(new java.awt.event.ActionListener() {
 			public void actionPerformed(java.awt.event.ActionEvent e) {
@@ -115,17 +100,4 @@ public class ClienteActionsCellEditor extends AbstractCellEditor implements Tabl
 		return null;
 	}
 
-	@Override
-	public boolean isCellEditable(EventObject e) {
-		return true;
-	}
-
-	private static JButton iconButton(ImageIcon ico, String tip) {
-		JButton b = new JButton(ico);
-		b.setToolTipText(tip);
-		b.setBorderPainted(false);
-		b.setFocusPainted(false);
-		b.setContentAreaFilled(false);
-		return b;
-	}
 }
