+++ renderer/ReservaActionsCellEditor.java
@@ -1,21 +1,11 @@
 package com.pinguela.rentexpres.desktop.renderer;
 
-import static com.pinguela.rentexpres.desktop.util.AppIcons.DELETE;
-import static com.pinguela.rentexpres.desktop.util.AppIcons.EDIT;
-import static com.pinguela.rentexpres.desktop.util.AppIcons.VIEW;
-
 import java.awt.Component;
-import java.awt.FlowLayout;
 import java.awt.Frame;
 import java.util.EventObject;
 
-import javax.swing.AbstractCellEditor;
-import javax.swing.ImageIcon;
-import javax.swing.JButton;
 import javax.swing.JOptionPane;
-import javax.swing.JPanel;
 import javax.swing.JTable;
-import javax.swing.table.TableCellEditor;
 
 import com.pinguela.rentexpres.desktop.dialog.ReservaDetailDialog;
 import com.pinguela.rentexpres.desktop.dialog.ReservaEditDialog;
@@ -25,20 +15,13 @@ import com.pinguela.rentexpres.exception.RentexpresException;
 import com.pinguela.rentexpres.model.ReservaDTO;
 import com.pinguela.rentexpres.service.ReservaService;
 
-public class ReservaActionsCellEditor extends AbstractCellEditor implements TableCellEditor {
-	private static final long serialVersionUID = 1L;
+public class ReservaActionsCellEditor extends AbstractActionsCellEditor {
+        private static final long serialVersionUID = 1L;
 
-	private final JPanel panel = new JPanel(new FlowLayout(FlowLayout.CENTER, 6, 0));
-	private final JButton btnVer = iconButton(VIEW, "Ver");
-	private final JButton btnEdit = iconButton(EDIT, "Editar");
-	private final JButton btnDel = iconButton(DELETE, "Borrar");
+        private ReservaDTO reservaActual;
 
-	private ReservaDTO reservaActual;
-
-	public ReservaActionsCellEditor(Frame frame, ReservaService service) {
-		panel.add(btnVer);
-		panel.add(btnEdit);
-		panel.add(btnDel);
+        public ReservaActionsCellEditor(Frame frame, ReservaService service) {
+                super();
 
 		btnVer.addActionListener(new java.awt.event.ActionListener() {
 			public void actionPerformed(java.awt.event.ActionEvent e) {
@@ -96,17 +79,4 @@ public class ReservaActionsCellEditor extends AbstractCellEditor implements Tabl
 		return null;
 	}
 
-	@Override
-	public boolean isCellEditable(EventObject e) {
-		return true;
-	}
-
-	private static JButton iconButton(ImageIcon icon, String tooltip) {
-		JButton b = new JButton(icon);
-		b.setToolTipText(tooltip);
-		b.setBorderPainted(false);
-		b.setFocusPainted(false);
-		b.setContentAreaFilled(false);
-		return b;
-	}
 }
