+++ renderer/VehiculoActionsCellEditor.java
@@ -1,23 +1,14 @@
 package com.pinguela.rentexpres.desktop.renderer;
 
-import static com.pinguela.rentexpres.desktop.util.AppIcons.DELETE;
-import static com.pinguela.rentexpres.desktop.util.AppIcons.EDIT;
-import static com.pinguela.rentexpres.desktop.util.AppIcons.VIEW;
-
 import java.awt.Component;
-import java.awt.FlowLayout;
 import java.awt.Frame;
-import java.util.EventObject;
 import java.util.List;
 
-import javax.swing.AbstractCellEditor;
-import javax.swing.ImageIcon;
-import javax.swing.JButton;
 import javax.swing.JOptionPane;
-import javax.swing.JPanel;
 import javax.swing.JTable;
 import javax.swing.SwingUtilities;
-import javax.swing.table.TableCellEditor;
+
+import com.pinguela.rentexpres.desktop.renderer.AbstractActionsCellEditor;
 
 import com.pinguela.rentexpres.desktop.controller.SearchVehiculoAction;
 import com.pinguela.rentexpres.desktop.dialog.VehiculoDetailDialog;
@@ -30,28 +21,20 @@ import com.pinguela.rentexpres.model.EstadoVehiculoDTO;
 import com.pinguela.rentexpres.model.VehiculoDTO;
 import com.pinguela.rentexpres.service.VehiculoService;
 
-public class VehiculoActionsCellEditor extends AbstractCellEditor implements TableCellEditor {
-	private static final long serialVersionUID = 1L;
+public class VehiculoActionsCellEditor extends AbstractActionsCellEditor {
+        private static final long serialVersionUID = 1L;
 
-	private final JPanel panel = new JPanel(new FlowLayout(FlowLayout.CENTER, 6, 0));
-	private final JButton btnVer = iconButton(VIEW, "Ver");
-	private final JButton btnEdit = iconButton(EDIT, "Editar");
-	private final JButton btnDel = iconButton(DELETE, "Borrar");
-
-	private final Frame frame;
-	private final VehiculoService service;
-	private final SearchVehiculoAction searchAction;
+        private final Frame frame;
+        private final VehiculoService service;
+        private final SearchVehiculoAction searchAction;
 
 	private VehiculoDTO vehiculoActual;
 
-	public VehiculoActionsCellEditor(JTable table, VehiculoService vehiculoService, SearchVehiculoAction searchAction) {
-		this.frame = (Frame) SwingUtilities.getWindowAncestor(table);
-		this.service = vehiculoService;
-		this.searchAction = searchAction;
-
-		panel.add(btnVer);
-		panel.add(btnEdit);
-		panel.add(btnDel);
+        public VehiculoActionsCellEditor(JTable table, VehiculoService vehiculoService, SearchVehiculoAction searchAction) {
+                super();
+                this.frame = (Frame) SwingUtilities.getWindowAncestor(table);
+                this.service = vehiculoService;
+                this.searchAction = searchAction;
 
 		// Acción Ver
 		btnVer.addActionListener(new java.awt.event.ActionListener() {
@@ -122,19 +105,7 @@ public class VehiculoActionsCellEditor extends AbstractCellEditor implements Tab
 		return null;
 	}
 
-	@Override
-	public boolean isCellEditable(EventObject e) {
-		return true;
-	}
 
-	private static JButton iconButton(ImageIcon ico, String tip) {
-		JButton b = new JButton(ico);
-		b.setToolTipText(tip);
-		b.setBorderPainted(false);
-		b.setFocusPainted(false);
-		b.setContentAreaFilled(false);
-		return b;
-	}
 
 	public SearchVehiculoAction getSearchAction() {
 		return searchAction;
